+++ packages/plugin/docs/examples/main.ts
@@ -1,6 +1,6 @@
 
-import { hideBin } from '@dovenv/core/utils'
 import { Dovenv }  from '@dovenv/core'
+import { hideBin } from '@dovenv/core/utils'
 
 import config from './config'