+++ packages/utils/src/md/main.example.ts
@@ -8,33 +8,44 @@ import {
 	getMD,
 	geMDTocString,
 	md2html,
+	html2terminal,
+	line,
+	color,
 } from '../main'
 
+const title         = ( t: string ) => console.log( `\n\n${line( { title: color.inverse( ` ${t} ` ) } )}\n\n` )
 const currDir       = await getCurrentDir( import.meta.url )
 const pkgDir        = joinPath( currDir, '..', '..' )
 const readmePath    = joinPath( pkgDir, 'README.md' )
 const readmeContent = await getMD( readmePath )
 
-console.log( 'Terminal markdown' )
+title( 'Terminal markdown' )
 console.log( await md2terminal(
 	readmePath,
-	{ renderer : {
+	{
+		renderer : {
 		// @ts-ignore
-		image : function ( _href, _title, text ) {
+			image : function ( _href, _title, text ) {
 
-			return text
+				return text
 
-		} } },
+			},
+		},
+	},
 ) )
 
-console.log( 'md to html' )
+title( 'md to html' )
 console.log( await md2html( readmeContent ) )
 
-console.log( 'Increment headers' )
+title( 'html to terminal' )
+console.log( await html2terminal( await md2html( readmeContent ) ) )
+
+title( 'Increment headers' )
 const cont = await incrementMdHeaders( readmeContent )
 console.log( await md2terminal( cont ), cont )
 
-console.log( 'Readme toc\n', await geMDTocString( {
+title( 'Readme toc' )
+console.log( await geMDTocString( {
 	input    : readmeContent,
 	title    : 'Table of contents',
 	removeH1 : true,