+++ packages/create/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dovenv
 
+## 0.0.12
+
+### Patch Changes
+
+- fix issues
+
 ## 0.0.11
 
 ### Patch Changes