+++ packages/plugin/ai/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dovenv/ai
 
+## 0.0.12
+
+### Patch Changes
+
+- fix issues
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.12
+
 ## 0.0.11
 
 ### Patch Changes