+++ packages/plugin/lint/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @dovenv/lint
 
+## 0.0.12
+
+### Patch Changes
+
+- fix issues
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.12
+  - @dovenv/eslint-config@0.0.12
+  - @dovenv/stylelint-config@0.0.12
+
 ## 0.0.11
 
 ### Patch Changes